error handling

+++ commands/update.go
@@ -2,6 +2,8 @@ package commands
 
 import (
 	"os"
+	"errors"
+	"fmt"
 
 	"com.github/axblade/playwright/utils"
 )
@@ -14,13 +16,12 @@ func (self *UpdateCommand) Execute() (err error) {
 	folders := self.SelectFolders()
 
 	rolesPath, err := self.ReadRolesPath()
+
 	if err != nil {
 		return err
 	}
 
-	updatePlaybookStructure(rolesPath, self.Command.PlaybookName, folders)
-
-	return nil
+	return updatePlaybookStructure(rolesPath, self.Command.PlaybookName, folders)
 }
 
 func updatePlaybookStructure(rolesPath string, name string, folders []string) (err error) {
@@ -28,22 +29,30 @@ func updatePlaybookStructure(rolesPath string, name string, folders []string) (e
 		rolesPath = utils.Concat(rolesPath, "/")
 	}
 
-	// ToDo: check if role exists
-
 	playbookPath := utils.Concat(rolesPath, name)
+
+	if !utils.FolderExists(playbookPath) {
+		return errors.New(fmt.Sprintf("Role %s does not exists", name))
+	}
 
 	if string(playbookPath[len(playbookPath)-1]) != "/" {
 		playbookPath = utils.Concat(playbookPath, "/")
 	}
 
 	for _, folder := range folders {
-		folderPath := utils.Concat(playbookPath, folder)
-		// ToDo: check if folder exists, otehrwise throw error
-		os.MkdirAll(folderPath, 0755)
+		if folder != "tasks" {
+			folderPath := utils.Concat(playbookPath, folder)
 
-		if folder != "files" && folder != "templates" {
-			filePath := utils.Concat(folderPath, "/main.yml")
-			os.Create(filePath)
+			if utils.FolderExists(folderPath) {
+				return errors.New(fmt.Sprintf("Folder %s already exists for role %s", folder, name))
+			}
+
+			os.MkdirAll(folderPath, 0755)
+
+			if folder != "files" && folder != "templates" {
+				filePath := utils.Concat(folderPath, "/main.yml")
+				os.Create(filePath)
+			}
 		}
 	}
